Updated sample and read me to reflect updates to iOS SDK and Apple Developer Portal
