--- notification_hub_rest/Sample/NotificationHubSample/Foundation.Notifications/Foundation.Notifications/Foundation_Notifications.h
@@ -1,20 +1,0 @@
-//
-//  Foundation_Notifications.h
-//  Foundation.Notifications
-//
-//  Created by Mike Parker on 27/04/2018.
-//  Copyright © 2018 mobcat. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-#import <CommonCrypto/CommonHMAC.h>
-
-//! Project version number for Foundation_Notifications.
-FOUNDATION_EXPORT double Foundation_NotificationsVersionNumber;
-
-//! Project version string for Foundation_Notifications.
-FOUNDATION_EXPORT const unsigned char Foundation_NotificationsVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Foundation_Notifications/PublicHeader.h>
-
-
