--- notification_hub_rest/Sample/NotificationHubSample/Foundation.Notifications/Foundation.Notifications/ProjectModule/module.modulemap
@@ -1,12 +1,0 @@
-//
-//  module.modulemap
-//  Foundation.Notifications
-//
-//  Created by Mike Parker on 27/04/2018.
-//  Copyright © 2018 mobcat. All rights reserved.
-//
-
-module FoundationNotificationsPrivate {
-    header "../Foundation_Notifications.h"
-    export *
-}
