--- notification_hub_rest/Sample/NotificationHubSample/NotificationHubSample/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  NotificationHubSample
-//
-//  Created by Mike Parker on 27/04/2018.
-//  Copyright © 2018 mobcat. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
