+++ notification_hub_rest/Sample/PushDemo/PushDemo/AppDelegate.swift
@@ -1,0 +1,109 @@
+//
+//  AppDelegate.swift
+//  PushDemo
+//
+//  Created by Mike Parker on 27/04/2018.
+//  Copyright © 2018 mobcat. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterDelegate {
+
+    var configValues: NSDictionary?
+    var notificationHubNamespace : String?
+    var notificationHubName : String?
+    var notificationHubKeyName : String?
+    var notificationHubKey : String?
+    var registrationService : NotificationRegistrationService?
+    let tags = ["12345"]
+    let genericTemplate = PushTemplate(withBody: "{\"aps\":{\"alert\":\"$(message)\"}}")
+    
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        
+        if let path = Bundle.main.path(forResource: "devsettings", ofType: "plist") {
+            if let configValues = NSDictionary(contentsOfFile: path) {
+                self.notificationHubNamespace = configValues["notificationHubNamespace"] as? String
+                self.notificationHubName = configValues["notificationHubName"] as? String
+                self.notificationHubKeyName = configValues["notificationHubKeyName"] as? String
+                self.notificationHubKey = configValues["notificationHubKey"] as? String
+            }
+        }
+            
+        if #available(iOS 10.0, *){
+            UNUserNotificationCenter.current().delegate = self
+            UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound, .badge]) {
+                (granted, error) in
+                    
+                if (granted)
+                {
+                    DispatchQueue.main.async {
+                        application.registerForRemoteNotifications()
+                    }
+                }
+            }
+        }
+        
+        return true
+    }
+    
+    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+        let installationId = (UIDevice.current.identifierForVendor?.description)!
+        let pushChannel = deviceToken.reduce("", {$0 + String(format: "%02X", $1)})
+
+        // Initialize the Notification Registration Service
+        self.registrationService = NotificationRegistrationService(
+            withInstallationId: installationId,
+            andPushChannel: pushChannel,
+            andHubNamespace: notificationHubNamespace!,
+            andHubName: notificationHubName!,
+            andKeyName: notificationHubKeyName!,
+            andKey: notificationHubKey!)
+    
+        // Call register passing in the tags and template parameters
+        self.registrationService!.register(withTags: tags, andTemplates: ["genericTemplate" : self.genericTemplate]) { (result) in
+            if !result {
+                print("Registration issue")
+            } else {
+                print("Registered")
+            }
+        }
+    }
+
+    func showAlert(withText text : String) {
+        let alertController = UIAlertController(title: "PushDemo", message: text, preferredStyle: UIAlertController.Style.alert)
+        alertController.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default,handler: nil))
+        let keyWindow = UIApplication.shared.windows.filter {$0.isKeyWindow}.first
+        keyWindow?.rootViewController?.present(alertController, animated: true, completion: nil)
+    }
+    
+    @available(iOS 10.0, *)
+    func userNotificationCenter(_ center: UNUserNotificationCenter,
+        willPresent notification: UNNotification,
+        withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+        showAlert(withText: notification.request.content.body)
+    }
+    
+    @available(iOS 10.0, *)
+    func userNotificationCenter(_ center: UNUserNotificationCenter,
+        didReceive response: UNNotificationResponse,
+        withCompletionHandler completionHandler: @escaping () -> Void) {
+        showAlert(withText: response.notification.request.content.body)
+    }
+
+    // MARK: UISceneSession Lifecycle
+
+    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
+        // Called when a new scene session is being created.
+        // Use this method to select a configuration to create the new scene with.
+        return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+    }
+
+    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
+        // Called when the user discards a scene session.
+        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
+        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
+    }
+}
