+++ notification_hub_rest/Sample/PushDemo/PushDemo/Assets.xcassets/AppIcon.appiconset/Contents.json
