+++ notification_hub_rest/Sample/PushDemo/PushDemo/Models/PushTemplate.swift
@@ -1,8 +1,8 @@
 //
 //  PushTemplate.swift
-//  NotificationHubSample
+//  PushDemo
 //
-//  Created by Mike Parker on 04/05/2018.
+//  Created by Mike Parker on 27/04/2018.
 //  Copyright © 2018 mobcat. All rights reserved.
 //
