+++ notification_hub_rest/Sample/PushDemo/PushDemo/Models/TokenData.swift
@@ -1,15 +1,15 @@
 //
 //  TokenData.swift
-//  NotificationHubSample
+//  PushDemo
 //
-//  Created by Mike Parker on 04/05/2018.
+//  Created by Mike Parker on 27/04/2018.
 //  Copyright © 2018 mobcat. All rights reserved.
 //
 
 import Foundation
 
 struct TokenData {
-    
+
     let token : String
     let expiration : Int
