+++ notification_hub_rest/Sample/PushDemo/PushDemo/PushDemo.entitlements
@@ -2,7 +2,7 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>IDEDidComputeMac32BitWarning</key>
-	<true/>
+	<key>aps-environment</key>
+	<string>development</string>
 </dict>
 </plist>
