+++ notification_hub_rest/Sample/PushDemo/PushDemo/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  PushDemo
+//
+//  Created by Mike Parker on 27/04/2018.
+//  Copyright © 2018 mobcat. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
